requirements.txt
tensorflow
keras
fastapi
uvicorn
python-multipart
pillow
google-cloud-aiplatform
numpy
requests
python-dotenv
